2017-05-13 update using the new algorith. Expect some deletions

+++ distros/p/perl/lib/h2xs.t
@@ -31,8 +31,28 @@ $ExtUtils::Manifest::Quiet=1;
 my $up = File::Spec->updir();
 
 my $extracted_program = '../utils/h2xs'; # unix, nt, ...
-if ($^O eq 'VMS') { $extracted_program = '[-.utils]h2xs.com'; }
-if ($^O eq 'MacOS') { $extracted_program = '::utils:h2xs'; }
+
+my $Is_VMS_traildot = 0;
+if ($^O eq 'VMS') {
+    $extracted_program = '[-.utils]h2xs.com';
+
+    # We have to know if VMS is in UNIX mode.  In UNIX mode, trailing dots
+    # should not be present.  There are actually two settings that control this.
+
+    $Is_VMS_traildot = 1;
+    my $unix_rpt = 0;
+    my $drop_dot = 0;
+    if (eval 'require VMS::Feature') {
+        $unix_rpt = VMS::Feature::current('filename_unix_report');
+        $drop_dot = VMS::Feature::current('readdir_dropdotnotype');
+    } else {
+        my $unix_report = $ENV{'DECC$FILENAME_UNIX_REPORT'} || '';
+        $unix_rpt = $unix_report =~ /^[ET1]/i; 
+        my $drop_dot_notype = $ENV{'DECC$READDIR_DROPDOTNOTYPE'} || '';
+        $drop_dot = $drop_dot_notype =~ /^[ET1]/i;
+    }
+    $Is_VMS_traildot = 0 if $drop_dot && unix_rpt;
+}
 if (!(-e $extracted_program)) {
     print "1..0 # Skip: $extracted_program was not built\n";
     exit 0;
@@ -44,21 +64,15 @@ if (!(-e $extracted_program)) {
 my $dupe = '2>&1';
 # ok on unix, nt, The extra \" are for VMS
 my $lib = '"-I../lib" "-I../../lib"';
-# The >&1 would create a file named &1 on MPW (STDERR && STDOUT are
-# already merged).
-if ($^O eq 'MacOS') {
-    $dupe = '';
-    # -x overcomes MPW $Config{startperl} anomaly
-    $lib = '-x -I::lib: -I:::lib:';
-}
 # $name should differ from system header file names and must
 # not already be found in the t/ subdirectory for perl.
 my $name = 'h2xst';
 my $header = "$name.h";
 my $thisversion = sprintf "%vd", $^V;
+$thisversion =~ s/^v//;
 
 # If this test has failed previously a copy may be left.
-rmtree($name);
+rmtree($name) if -e $name;
 
 my @tests = (
 "-f -n $name", $], <<"EOXSFILES",
@@ -145,8 +159,10 @@ for (my $i = $#tests; $i > 0; $i-=3) {
   # 1 test for running it, 1 test for the expected result, and 1 for each file
   # plus 1 to open and 1 to check for the use in lib/$name.pm and Makefile.PL
   # And 1 more for our check for "bonus" files, 2 more for ExtUtil::Manifest.
+  # And 1 more to examine const-c.inc contents in tests that use $header.
   # use the () to force list context and hence count the number of matches.
   $total_tests += 9 + (() = $tests[$i] =~ /(Writing)/sg);
+  $total_tests++ if $tests[$i-2] =~ / \Q$header\E$/;
 }
 
 plan tests => $total_tests;
@@ -155,6 +171,8 @@ ok (open (HEADER, ">$header"), "open '$header'");
 print HEADER <<HEADER or die $!;
 #define Camel 2
 #define Dromedary 1
+#define Bactrian /* empty */
+#define Bactrian2
 HEADER
 ok (close (HEADER), "close '$header'");
 
@@ -167,11 +185,6 @@ while (my ($args, $version, $expectation) = splice @tests, 0, 3) {
   cmp_ok ($?, "==", 0, "running $prog ");
   $result = join("",@result);
 
-  # accomodate MPW # comment character prependage
-  if ($^O eq 'MacOS') {
-    $result =~ s/#\s*//gs;
-  }
-
   #print "# expectation is >$expectation<\n";
   #print "# result is >$result<\n";
   # Was the output the list of files that were expected?
@@ -181,12 +194,10 @@ while (my ($args, $version, $expectation) = splice @tests, 0, 3) {
   find (sub {$got{$File::Find::name}++ unless -d $_}, $name);
 
   foreach ($expectation =~ /Writing\s+(\S+)/gm) {
-    if ($^O eq 'MacOS') {
-      $_ = ':' . join(':',split(/\//,$_));
-      $_ =~ s/$name:t:1.t/$name:t\/1.t/; # is this an h2xs bug?
-    }
     if ($^O eq 'VMS') {
-      $_ .= '.' unless $_ =~ m/\./;
+      if ($Is_VMS_traildot) {
+          $_ .= '.' unless $_ =~ m/\./;
+      }
       $_ = lc($_) unless exists $got{$_};
     }
     ok (-e $_, "check for $_") and delete $got{$_};
@@ -208,6 +219,23 @@ while (my ($args, $version, $expectation) = splice @tests, 0, 3) {
   pop @INC;
   chdir ($up) or die "chdir $up failed: $!";
  
+  if ($args =~ / \Q$header\E$/) {
+    my $const_c = File::Spec->catfile($name, 'fallback', 'const-c.inc');
+    my ($found, $diag);
+    if (!open FILE, '<', $const_c) {
+      $diag = "can't open $const_c: $!";
+    }
+    else {
+      while (<FILE>) {
+        next unless /\b Bactrian 2? \b/x;
+        $found = 1;
+        last;
+      }
+    }
+    ok (!$found, "generated $const_c has no Bactrian(2)");
+    diag ($diag) if defined $diag;
+  }
+
   foreach my $leaf (File::Spec->catfile('lib', "$name.pm"), 'Makefile.PL') {
     my $file = File::Spec->catfile($name, $leaf);
     if (ok (open (FILE, $file), "open $file")) {
